Move zip.ts to /common
