+++ extensions/ql-vscode/src/common/zip.ts
